Review feedback changes
Fix remaining "run.sh" instances.

Remove unmotivated newline from run.ps1.

In "-h" output, don't show "pwsh". It's not always required, and if the user has gotten this far, they're already using the right version of PowerShell.

Add Prerequisites section to eng/README.md and clarify that run.ps1 is cross-platform. Add hint to the utilities.ps1 PowerShell 6+ requirement detector that suggests "pwsh" prefix might be missing.

Don't include the path to run.ps1 in the list of possible tools. It's clutter (hard to read) and the user is better off repeating the last command and adding the tool they want. We don't know if a "pwsh" prefix was necessary, for example.