Use SBOM template and build retention from Arcade
Always generate an SBOM. The SBOM is meant to contain the version, so this commit also adds a releaseVersion parameter.

Retain the build forever when this build is a versioned release. This also relies on the new releaseVersion parameter.

A change in go-infra release automation will be required to pass in releaseVersion.