ed25519: check for boring enabled, and rand at the same time (#1786)
When compiling upstream projects (in particular
https://github.com/nats-io/nkeys) it is possible to reach Unreachable
code path.

It seems much safer to check for rand reader at the same time as
boring.Enabled and boring.SupportsEd25519 as otherwise automatic
fallback to the other implementation does not happen.

If the intention is to automatically allow using fallback, as the code
suggests to be the case.

Reported-by: @javacruft